+++ client/templates/logic/bugReport.js
@@ -23,11 +23,14 @@ function sendBugReport(br) {
 Template.account.events({
     'click #removeAccount': function(e) {
         e.preventDefault();
-        Meteor.call('removeAccount', Meteor.userId(), Meteor.userId(), function(err, result) {
-            if (result)
-                alert("Done.");
-            if (err)
-                alert("Error.");
-        });
+        var username = prompt("Please, write your username: ");
+        if (username !== null && username.replace(/\s/g, '').toLowerCase() === Meteor.user().profile.name.replace(/\s/g, '').toLowerCase()) {
+            Meteor.call('removeAccount', Meteor.userId(), Meteor.userId(), function(err, result) {
+                if (result)
+                    alert("Done.");
+                if (err)
+                    alert("Error.");
+            });
+        }
     }
 });