+++ client/templates/logic/s-movie.js
@@ -164,4 +164,5 @@ function resetVariables() {
     arrayResultGenre = [];
     arrayResultKeyword = [];
     Session.set('numberOfResults', 0);
+    Session.set("moreResults", false);
 }